Improve implementation of <Stack/> (formerly <ButtonStack/>)
Unfortunately, to have behavior like <xul:deck>, the only way I can figure out
is just to use the XUL box model. :(
