+++ src/webextension/widgets/password-input.js
@@ -6,7 +6,7 @@ import { Localized } from "fluent-react";
 import PropTypes from "prop-types";
 import React from "react";
 
-import ButtonStack from "./button-stack";
+import Stack from "./stack";
 
 import styles from "./input.css";
 
@@ -57,7 +57,7 @@ export default class PasswordInput extends React.Component {
         <input className={monospace ? styles.monospace : ""} disabled={disabled}
                type={showPassword ? "text" : "password"}
                ref={(element) => this.inputElement = element} {...props}/>
-        <ButtonStack selectedIndex={selectedIndex}>
+        <Stack stretch selectedIndex={selectedIndex}>
           <Localized id="password-input-show">
             <button className={styles.showBtn} type="button" title="sHOw" disabled={disabled}
                     onClick={() => this.showPassword(true)}></button>
@@ -66,7 +66,7 @@ export default class PasswordInput extends React.Component {
             <button className={styles.hideBtn} type="button" title="hIDe" disabled={disabled}
                     onClick={() => this.showPassword(false)}></button>
           </Localized>
-        </ButtonStack>
+        </Stack>
       </div>
     );
   }
