+++ src/webextension/widgets/stack.css
@@ -2,17 +2,20 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-/* This should work like <xul:deck>, but it has some significant limitations
-   (mainly that every element should be the same height). Perhaps there's a
-   better way to do all this... */
+/* XXX: Sadly, I can't come up with a better way to make something that works
+   like a <xul:deck> than by just using XUL's box model... */
 
 .stack {
-  display: flex;
-  flex-direction: column;
+  display: -moz-stack;
+}
+
+.stretch {
+  height: 100%;
+  width: 100%;
 }
 
 .stack-item {
-  height: 0;
+  display: -moz-box;
   visibility: hidden;
 }
 
@@ -21,5 +24,5 @@
 }
 
 .stack-item > * {
-  width: 100%;
+  -moz-box-flex: 1;
 }
